refactor(fw): #13 extract relay wire codec to pure net::wire + host byte-compat guard
Move the SMOLv1 relay-family wire codec (encode/parse for RELAY, RELAYACK, RELAY2,
RELAYACK2, BATT2/GRID2 via encode_dl/parse_dl, synth_origin_mac) + its fixed-width
ASCII field helpers (write_u5/u10, parse_id/u5/u10) + the tag prefixes/RELAY_CHUNK/
BATT_PAYLOAD_MAX out of net/mode.rs into a PURE, no-HAL-deps net/wire.rs — mirroring
the net/flood.rs split so the frame formats are HOST-unit-testable off-target. Pure
move, byte-identical: mode.rs re-exports via `use crate::net::wire::*` so every call
site is unchanged, and the espnow,cast,io release ELF is byte-size-identical to the
pre-refactor build.

New host guard experiments/relay_compat `#[path]`-includes the REAL wire.rs and asserts
the mixed-fleet / rolling-upgrade byte contract BOTH directions — the permanent
regression the bench-cornered C0 investigation asked for, and the checkpoint the #124
UP2 migration must preserve/replace:
  1. new-code encode_relay(H=1) -> GOLDEN bytes + parses under a VENDORED pre-#13
     parser (a #13 leaf's H=1 RELAY is accepted by an old gateway), incl. a max fragment;
  2. an old-format RELAYACK parses under the NEW matcher (the #13 leaf DOES recognise a
     plain RELAYACK — settling the C0 ACK-recognition question in code);
  3. RELAY2/RELAYACK2/BATT2 round-trips + DISAMBIGUATION (a plain RELAY never classifies
     as RELAY2 and vice-versa; RELAYACK never as RELAYACK2; BATT2 never as GRID2) — the
     no-flag-day guarantee, executable.

Verified: clippy -D warnings clean on default/espnow/espnow,cast,io/espnow,wled/
espnow,mesh-test; release links clean on espnow + espnow,cast,io; relay_compat ALL PASS.

Refs #13.
